test: split vsock host process behavior tests
