+++ crates/vsock-host/src/tests/mod.rs
@@ -3,3 +3,4 @@ pub(crate) mod support;
 mod command;
 mod connection;
 mod file;
+mod process;
